refactor: week get 리팩토링
