+++ src/week/week.repository.ts
@@ -4,9 +4,8 @@ import {
   InternalServerErrorException,
 } from '@nestjs/common';
 import { SemesterRepository } from 'src/semester/semester.repository';
-import { isBetweenWeek } from 'src/utils/isBetweenWeek';
 import { DataSource, Repository } from 'typeorm';
-import { CreateWeekDto, GetWeekDto, UpdateWeekDto } from './dto/week.dto';
+import { CreateWeekDto, UpdateWeekDto } from './dto/week.dto';
 import { week } from './week.entity';
 
 @Injectable()
@@ -72,41 +71,6 @@ export class WeekRepository extends Repository<week> {
     return {
       message: '업데이트 성공',
       data: await this.findOneBy({ semesterId }),
-    };
-  }
-
-  async getWeek(getWeekDto: GetWeekDto) {
-    const { date } = getWeekDto;
-
-    const [weeks] = await this.find({
-      order: {
-        id: 'DESC',
-      },
-      take: 1,
-    });
-
-    const weekObject = {
-      week1: weeks.week1,
-      week2: weeks.week2,
-      week3: weeks.week3,
-      week4: weeks.week4,
-      midterm: weeks.midterm,
-      week5: weeks.week5,
-      week6: weeks.week6,
-      week7: weeks.week7,
-      week8: weeks.week8,
-    };
-
-    for (const elem in weekObject) {
-      if (isBetweenWeek(date, weekObject[elem]) == true) {
-        return {
-          week: elem,
-        };
-      }
-    }
-
-    return {
-      message: '해당 주차는 존재하지 않습니다.',
     };
   }
 }
